✅ Palindrome Number | Easy | C++ | Runtime: 0 ms | Memory: 7.7 MB | 2026-09-19

+++ leetcode/easy/0009-palindrome-number/solution.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-        double rev_num=0;
+        long long rev_num=0;
     int num=x;
     int last_digit;
     while(num>0){
